tar i bruk samme bigquery setup for okr fortelling, setter felles tittel på siden med datostempel
